# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "L.U.C.I.A. API",
    "description": "Logic & Ultimate Chess Insight Algorithm",
    "version": "0.2.5"
  },
  "paths": {
...
    "/boards": {
      "get": {
        "tags": [
          "boards"
        ],
        "summary": "List Boards",
        "operationId": "list_boards_boards_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardSummary"
                  },
                  "title": "Response List Boards Boards Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Create Board",
        "operationId": "create_board_boards_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/boards/{board_id}": {
      "get": {
        "tags": [
          "boards"
        ],
        "summary": "Get Board",
        "operationId": "get_board_boards__board_id__get",
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Board Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "boards"
        ],
        "summary": "Update Board",
        "operationId": "update_board_boards__board_id__put",
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Board Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "boards"
        ],
        "summary": "Delete Board",
        "operationId": "delete_board_boards__board_id__delete",
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Board Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/boards/{board_id}/undo": {
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Undo Board",
        "description": "Devuelve el tablero a su estado anterior (RF-6.8).\n\n409 y no 400 cuando no hay nada que deshacer: la petición es correcta, es\nel estado del tablero el que no la admite.",
        "operationId": "undo_board_boards__board_id__undo_post",
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Board Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/boards/{board_id}/redo": {
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Redo Board",
        "operationId": "redo_board_boards__board_id__redo_post",
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Board Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/boards/{board_id}/own-game": {
      "put": {
        "tags": [
          "boards"
        ],
        "summary": "Publish Own Game",
        "description": "Publica el tablero en el historial como partida propia (RF-6.5).\n\nA partir de aquí cuenta en el dashboard, en los patrones y en los filtros\ndel listado como cualquier otra partida. La misma llamada sirve para\ncorregir los datos y para poner al día las jugadas después de deshacer o\nrehacer: reescribe la partida publicada en vez de crear otra.",
        "operationId": "publish_own_game_boards__board_id__own_game_put",
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Board Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnGamePublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "boards"
        ],
        "summary": "Withdraw Own Game",
        "description": "Retira la marca de partida propia: el tablero deja de contar y su\npartida se va del historial (RF-6.5). El tablero se queda como estaba.",
        "operationId": "withdraw_own_game_boards__board_id__own_game_delete",
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Board Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/boards/{board_id}/analysis": {
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Analyze Board",
        "description": "Encola el análisis completo de la línea principal del tablero (RF-6.9).\n\nDevuelve en cuanto está en la cola, con el id que el front usa para\nseguir el progreso por `WS /ws/analysis/{id}`, igual que una partida. El\nanálisis **no cuenta en las estadísticas** salvo que el tablero esté\npublicado como partida propia (RF-6.5), que es cuando nace con `game_id`.",
        "operationId": "analyze_board_boards__board_id__analysis_post",
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Board Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardAnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoardCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "root_fen": {
            "type": "string",
            "title": "Root Fen",
            "default": "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1"
          },
          "tree_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Tree Json"
          },
          "tags_json": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags Json"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "BoardCreate"
      },
      "BoardDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "tags_json": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags Json"
          },
          "root_fen": {
            "type": "string",
            "title": "Root Fen"
          },
          "is_own_game": {
            "type": "boolean",
            "title": "Is Own Game"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "tree_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Tree Json"
          },
          "can_undo": {
            "type": "boolean",
            "title": "Can Undo"
          },
          "can_redo": {
            "type": "boolean",
            "title": "Can Redo"
          },
          "own_game": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnGameLink"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "tags_json",
          "root_fen",
          "is_own_game",
          "created_at",
          "updated_at",
          "tree_json",
          "can_undo",
          "can_redo"
        ],
        "title": "BoardDetail"
      },
      "BoardSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "tags_json": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags Json"
          },
          "root_fen": {
            "type": "string",
            "title": "Root Fen"
          },
          "is_own_game": {
            "type": "boolean",
            "title": "Is Own Game"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "tags_json",
          "root_fen",
          "is_own_game",
          "created_at",
          "updated_at"
        ],
        "title": "BoardSummary",
        "description": "Lo que hace falta para listar; sin el árbol, que puede ser grande."
      },
      "BoardUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "root_fen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Fen"
          },
          "tree_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tree Json"
          },
          "tags_json": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags Json"
          },
          "pgn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pgn"
          }
        },
        "type": "object",
        "title": "BoardUpdate",
        "description": "Todos los campos son opcionales: la UI autoguarda solo lo que cambió."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OwnGamePublishRequest": {
        "properties": {
          "player_color": {
            "type": "string",
            "enum": [
              "white",
              "black"
            ],
            "title": "Player Color"
          },
          "opponent_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Opponent Name"
          },
          "result": {
            "type": "string",
            "enum": [
              "win",
              "draw",
              "loss"
            ],
            "title": "Result"
          },
          "played_on": {
            "type": "string",
            "format": "date",
            "title": "Played On"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "pgn": {
            "type": "string",
            "minLength": 1,
            "title": "Pgn"
          }
        },
        "type": "object",
        "required": [
          "player_color",
          "opponent_name",
          "result",
          "played_on",
          "pgn"
        ],
        "title": "OwnGamePublishRequest",
        "description": "Marcar el tablero como partida propia (RF-6.5): los datos que el\ntablero no tiene, más las jugadas que sí."
      },
...
    }
  }
}